fix: Resolve TypeError in VisualizationWidget paintEvent
Corrected a TypeError that occurred during the drawing of history bar shapes (circles) in the VisualizationWidget. The y-coordinate for the QRect constructor was being passed as a float value in some cases.

The fix involves explicitly casting the y-coordinate calculation to an integer for both Player 1 and Player 2 history drawing loops, ensuring that all arguments to QRect conform to the expected types. This prevents the application from crashing during the paintEvent.

+++ advanced_ipd_gui_v6.py
@@ -693,7 +693,7 @@ class VisualizationWidget(QWidget):
                 painter.setBrush(QBrush(QColor("#4CAF50")))
                 painter.setPen(QPen(QColor("#388E3C")))
                 side = min(int(draw_width), self.bar_height)
-                circle_rect = QRect(int(x + (draw_width - side) / 2), y_p1 + (self.bar_height - side) / 2, side, side)
+                circle_rect = QRect(int(x + (draw_width - side) / 2), int(y_p1 + (self.bar_height - side) / 2), side, side)
                 painter.drawEllipse(circle_rect)
             else: # DEFECT
                 painter.setBrush(QBrush(QColor("#FF9800")))
@@ -718,7 +718,7 @@ class VisualizationWidget(QWidget):
                 painter.setBrush(QBrush(QColor("#4CAF50")))
                 painter.setPen(QPen(QColor("#388E3C")))
                 side = min(int(draw_width), self.bar_height)
-                circle_rect = QRect(int(x + (draw_width - side) / 2), y_p2 + (self.bar_height - side) / 2, side, side)
+                circle_rect = QRect(int(x + (draw_width - side) / 2), int(y_p2 + (self.bar_height - side) / 2), side, side)
                 painter.drawEllipse(circle_rect)
             else: # DEFECT
                 painter.setBrush(QBrush(QColor("#FF9800")))
